Fix SQL function return type

+++ src/db/sql/functions/helper/fill_polygon_h3_3.sql
@@ -1,6 +1,6 @@
 DROP FUNCTION IF EXISTS basic.fill_polygon_h3_3;
 CREATE OR REPLACE FUNCTION basic.fill_polygon_h3_3(geom geometry)
-RETURNS TABLE (h3_index h3index, h3_short smallint, h3_boundary geometry(linestring, 4326), h3_geom geometry(polygon, 4326))
+RETURNS TABLE (h3_index h3index, h3_short int, h3_boundary geometry(linestring, 4326), h3_geom geometry(polygon, 4326))
 LANGUAGE plpgsql
 AS $function$
 BEGIN
